requirements.txt
google-adk>=0.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
mcp>=1.0.0
pydantic>=2.7.0
google-generativeai>=0.7.0
